allow for numba-less installation and compatibility with new numba-cuda
